Extract //cuttlefish/host/commands/run_cvd/launch:vhost_device_vsock
https://bazel.build/docs/bazel-and-cpp
