+++ base/cvd/cuttlefish/host/commands/run_cvd/BUILD.bazel
@@ -78,6 +78,7 @@ cc_binary(
         "//cuttlefish/host/commands/run_cvd/launch:tombstone_receiver",
         "//cuttlefish/host/commands/run_cvd/launch:uwb_connector",
         "//cuttlefish/host/commands/run_cvd/launch:vhal_proxy_server",
+        "//cuttlefish/host/commands/run_cvd/launch:vhost_device_vsock",
         "//cuttlefish/host/commands/run_cvd/launch:webrtc_controller",
         "//cuttlefish/host/commands/run_cvd/launch:wmediumd_server",
         "//cuttlefish/host/libs/command_util",
