+++ base/cvd/cuttlefish/host/commands/run_cvd/launch/vhost_device_vsock.cpp
@@ -25,7 +25,6 @@
 #include "cuttlefish/common/libs/utils/files.h"
 #include "cuttlefish/common/libs/utils/known_paths.h"
 #include "cuttlefish/common/libs/utils/result.h"
-#include "cuttlefish/host/commands/run_cvd/launch/launch.h"
 #include "cuttlefish/host/commands/run_cvd/launch/log_tee_creator.h"
 #include "cuttlefish/host/libs/config/command_source.h"
 #include "cuttlefish/host/libs/config/cuttlefish_config.h"
@@ -33,6 +32,31 @@
 #include "cuttlefish/host/libs/vm_manager/vm_manager.h"
 
 namespace cuttlefish {
+namespace {
+
+class VhostDeviceVsock : public vm_manager::VmmDependencyCommand {
+ public:
+  INJECT(VhostDeviceVsock(LogTeeCreator& log_tee,
+                          const CuttlefishConfig::InstanceSpecific& instance,
+                          const CuttlefishConfig& cfconfig));
+
+  // CommandSource
+  Result<std::vector<MonitorCommand>> Commands() override;
+
+  // SetupFeature
+  std::string Name() const override;
+  bool Enabled() const override;
+
+  Result<void> WaitForAvailability() const override;
+
+ private:
+  std::unordered_set<SetupFeature*> Dependencies() const override { return {}; }
+  Result<void> ResultSetup() override { return {}; }
+
+  LogTeeCreator& log_tee_;
+  const CuttlefishConfig::InstanceSpecific& instance_;
+  const CuttlefishConfig& cfconfig_;
+};
 
 VhostDeviceVsock::VhostDeviceVsock(
     LogTeeCreator& log_tee, const CuttlefishConfig::InstanceSpecific& instance,
@@ -89,6 +113,8 @@ Result<void> VhostDeviceVsock::WaitForAvailability() const {
   }
   return {};
 }
+
+}  // namespace
 
 fruit::Component<fruit::Required<const CuttlefishConfig, LogTeeCreator,
                                  const CuttlefishConfig::InstanceSpecific>>
