+++ base/cvd/cuttlefish/host/commands/run_cvd/launch/vhost_device_vsock.h
@@ -15,44 +15,15 @@
 
 #pragma once
 
-#include <string>
-#include <unordered_set>
-#include <utility>
-#include <vector>
-
 #include <fruit/fruit.h>
 
-#include "cuttlefish/common/libs/utils/files.h"
-#include "cuttlefish/common/libs/utils/result.h"
 #include "cuttlefish/host/commands/run_cvd/launch/log_tee_creator.h"
-#include "cuttlefish/host/libs/config/command_source.h"
 #include "cuttlefish/host/libs/config/cuttlefish_config.h"
-#include "cuttlefish/host/libs/vm_manager/vm_manager.h"
 
 namespace cuttlefish {
 
-class VhostDeviceVsock : public vm_manager::VmmDependencyCommand {
- public:
-  INJECT(VhostDeviceVsock(LogTeeCreator& log_tee,
-                          const CuttlefishConfig::InstanceSpecific& instance,
-                          const CuttlefishConfig& cfconfig));
-
-  // CommandSource
-  Result<std::vector<MonitorCommand>> Commands() override;
-
-  // SetupFeature
-  std::string Name() const override;
-  bool Enabled() const override;
-
-  Result<void> WaitForAvailability() const override;
-
- private:
-  std::unordered_set<SetupFeature*> Dependencies() const override { return {}; }
-  Result<void> ResultSetup() override { return {}; }
-
-  LogTeeCreator& log_tee_;
-  const CuttlefishConfig::InstanceSpecific& instance_;
-  const CuttlefishConfig& cfconfig_;
-};
+fruit::Component<fruit::Required<const CuttlefishConfig, LogTeeCreator,
+                                 const CuttlefishConfig::InstanceSpecific>>
+VhostDeviceVsockComponent();
 
 }  // namespace cuttlefish
